requirements.txt
# Core Django
Django==5.0.1

# Web Server
whitenoise==6.6.0

# CORS
django-cors-headers==4.3.1

# Forms & Validation
django-crispy-forms==2.1
crispy-tailwind==1.0.3

# API (pour les intégrations)
djangorestframework==3.14.0

# Utilities
python-dotenv==1.0.0
